systems vector link

+++ SBX/Main.cpp
@@ -3,7 +3,7 @@
 
 int main()
 {
-	std::srand( time( 0 ) );
+	std::srand( ( uint )time( 0 ) );
 	std::unique_ptr<Sandbox> sandbox = std::make_unique<Sandbox>();
 
 	LOG_ERROR( "Press any key to end the app" );
